Add files via upload
The title of this project is AUTOML VS CUSTOM TRAINED models. I have trained models on AUTOKERAS and trained my own custom models and have done a comparative analysis of both. My custom trained models were able to beat the performance of AUTOKERAS for image datasets. However, the analysis shows that autokeras performs really well when the data is tabular (regression problem). A report is also attached in the end which shows the comparative analysis with metrics such as accuracy, model loss, no. iterations (epochs) and time taken by each model to train.